+++ changelog.md
@@ -4,7 +4,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
-## [1.1.0] - 2022-03-12 
+## [1.1.0] - 2022-03-28
 
 ### Added